Update Azure modules to comply with AIP-21
