+++ airflow/contrib/hooks/azure_container_instance_hook.py
@@ -15,20 +15,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""
-This module is deprecated.
-Please use :mod:`airflow.providers.microsoft.azure.hooks.azure_container_instance`.
-"""
+"""This module is deprecated. Please use :mod:`airflow.providers.microsoft.azure.hooks.container_instance`."""
 
 import warnings
 
-from airflow.providers.microsoft.azure.hooks.azure_container_instance import (  # noqa
-    AzureContainerInstanceHook,
-)
+from airflow.providers.microsoft.azure.hooks.container_instance import AzureContainerInstanceHook  # noqa
 
 warnings.warn(
-    "This module is deprecated. "
-    "Please use `airflow.providers.microsoft.azure.hooks.azure_container_instance`.",
+    "This module is deprecated. " "Please use `airflow.providers.microsoft.azure.hooks.container_instance`.",
     DeprecationWarning,
     stacklevel=2,
 )
