+++ airflow/contrib/hooks/azure_container_registry_hook.py
@@ -15,20 +15,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""
-This module is deprecated.
-Please use `airflow.providers.microsoft.azure.hooks.azure_container_registry`.
-"""
+"""This module is deprecated. Please use `airflow.providers.microsoft.azure.hooks.container_registry`."""
 
 import warnings
 
-from airflow.providers.microsoft.azure.hooks.azure_container_registry import (  # noqa
-    AzureContainerRegistryHook,
-)
+from airflow.providers.microsoft.azure.hooks.container_registry import AzureContainerRegistryHook  # noqa
 
 warnings.warn(
-    "This module is deprecated. "
-    "Please use `airflow.providers.microsoft.azure.hooks.azure_container_registry`.",
+    "This module is deprecated. " "Please use `airflow.providers.microsoft.azure.hooks.container_registry`.",
     DeprecationWarning,
     stacklevel=2,
 )
