+++ airflow/contrib/hooks/azure_data_lake_hook.py
@@ -15,14 +15,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""This module is deprecated. Please use :mod:`airflow.providers.microsoft.azure.hooks.azure_data_lake`."""
+"""This module is deprecated. Please use :mod:`airflow.providers.microsoft.azure.hooks.data_lake`."""
 
 import warnings
 
-from airflow.providers.microsoft.azure.hooks.azure_data_lake import AzureDataLakeHook  # noqa
+from airflow.providers.microsoft.azure.hooks.data_lake import AzureDataLakeHook  # noqa
 
 warnings.warn(
-    "This module is deprecated. Please use `airflow.providers.microsoft.azure.hooks.azure_data_lake`.",
+    "This module is deprecated. Please use `airflow.providers.microsoft.azure.hooks.data_lake`.",
     DeprecationWarning,
     stacklevel=2,
 )
