+++ airflow/contrib/hooks/azure_fileshare_hook.py
@@ -15,14 +15,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""This module is deprecated. Please use :mod:`airflow.providers.microsoft.azure.hooks.azure_fileshare`."""
+"""This module is deprecated. Please use :mod:`airflow.providers.microsoft.azure.hooks.fileshare`."""
 
 import warnings
 
-from airflow.providers.microsoft.azure.hooks.azure_fileshare import AzureFileShareHook  # noqa
+from airflow.providers.microsoft.azure.hooks.fileshare import AzureFileShareHook  # noqa
 
 warnings.warn(
-    "This module is deprecated. Please use `airflow.providers.microsoft.azure.hooks.azure_fileshare`.",
+    "This module is deprecated. Please use `airflow.providers.microsoft.azure.hooks.fileshare`.",
     DeprecationWarning,
     stacklevel=2,
 )
