+++ airflow/contrib/operators/azure_cosmos_operator.py
@@ -15,14 +15,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""This module is deprecated. Please use :mod:`airflow.providers.microsoft.azure.operators.azure_cosmos`."""
+"""This module is deprecated. Please use :mod:`airflow.providers.microsoft.azure.operators.cosmos`."""
 
 import warnings
 
-from airflow.providers.microsoft.azure.operators.azure_cosmos import AzureCosmosInsertDocumentOperator  # noqa
+from airflow.providers.microsoft.azure.operators.cosmos import AzureCosmosInsertDocumentOperator  # noqa
 
 warnings.warn(
-    "This module is deprecated. Please use `airflow.providers.microsoft.azure.operators.azure_cosmos`.",
+    "This module is deprecated. Please use `airflow.providers.microsoft.azure.operators.cosmos`.",
     DeprecationWarning,
     stacklevel=2,
 )
