+++ airflow/contrib/sensors/azure_cosmos_sensor.py
@@ -15,14 +15,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""This module is deprecated. Please use :mod:`airflow.providers.microsoft.azure.sensors.azure_cosmos`."""
+"""This module is deprecated. Please use :mod:`airflow.providers.microsoft.azure.sensors.cosmos`."""
 
 import warnings
 
-from airflow.providers.microsoft.azure.sensors.azure_cosmos import AzureCosmosDocumentSensor  # noqa
+from airflow.providers.microsoft.azure.sensors.cosmos import AzureCosmosDocumentSensor  # noqa
 
 warnings.warn(
-    "This module is deprecated. Please use `airflow.providers.microsoft.azure.sensors.azure_cosmos`.",
+    "This module is deprecated. Please use `airflow.providers.microsoft.azure.sensors.cosmos`.",
     DeprecationWarning,
     stacklevel=2,
 )
