+++ airflow/providers/google/cloud/transfers/adls_to_gcs.py
@@ -25,7 +25,7 @@ from tempfile import NamedTemporaryFile
 from typing import Optional, Sequence, Union
 
 from airflow.providers.google.cloud.hooks.gcs import GCSHook, _parse_gcs_url
-from airflow.providers.microsoft.azure.hooks.azure_data_lake import AzureDataLakeHook
+from airflow.providers.microsoft.azure.hooks.data_lake import AzureDataLakeHook
 from airflow.providers.microsoft.azure.operators.adls import ADLSListOperator
