+++ airflow/providers/google/cloud/transfers/azure_fileshare_to_gcs.py
@@ -22,7 +22,7 @@ from typing import Iterable, Optional, Sequence, Union
 from airflow import AirflowException
 from airflow.models import BaseOperator
 from airflow.providers.google.cloud.hooks.gcs import GCSHook, _parse_gcs_url, gcs_object_is_directory
-from airflow.providers.microsoft.azure.hooks.azure_fileshare import AzureFileShareHook
+from airflow.providers.microsoft.azure.hooks.fileshare import AzureFileShareHook
 
 
 class AzureFileShareToGCSOperator(BaseOperator):
