+++ airflow/providers/microsoft/azure/example_dags/example_azure_container_instances.py
@@ -21,9 +21,7 @@ This is an example dag for using the AzureContainerInstancesOperator.
 from datetime import datetime, timedelta
 
 from airflow import DAG
-from airflow.providers.microsoft.azure.operators.azure_container_instances import (
-    AzureContainerInstancesOperator,
-)
+from airflow.providers.microsoft.azure.operators.container_instances import AzureContainerInstancesOperator
 
 with DAG(
     dag_id='aci_example',
