+++ airflow/providers/microsoft/azure/hooks/azure_container_instance.py
@@ -1,4 +1,3 @@
-#
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -15,143 +14,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-#
+"""This module is deprecated. Please use :mod:`airflow.providers.microsoft.azure.hooks.container_instance`."""
 
 import warnings
-from typing import Any
 
-from azure.mgmt.containerinstance import ContainerInstanceManagementClient
-from azure.mgmt.containerinstance.models import ContainerGroup
+from airflow.providers.microsoft.azure.hooks.container_instance import AzureContainerInstanceHook  # noqa
 
-from airflow.providers.microsoft.azure.hooks.base_azure import AzureBaseHook
-
-
-class AzureContainerInstanceHook(AzureBaseHook):
-    """
-    A hook to communicate with Azure Container Instances.
-
-    This hook requires a service principal in order to work.
-    After creating this service principal
-    (Azure Active Directory/App Registrations), you need to fill in the
-    client_id (Application ID) as login, the generated password as password,
-    and tenantId and subscriptionId in the extra's field as a json.
-
-    :param conn_id: :ref:`Azure connection id<howto/connection:azure>` of
-        a service principal which will be used to start the container instance.
-    :type azure_conn_id: str
-    """
-
-    conn_name_attr = 'azure_conn_id'
-    default_conn_name = 'azure_default'
-    conn_type = 'azure_container_instance'
-    hook_name = 'Azure Container Instance'
-
-    def __init__(self, *args, **kwargs) -> None:
-        super().__init__(sdk_client=ContainerInstanceManagementClient, *args, **kwargs)
-        self.connection = self.get_conn()
-
-    def create_or_update(self, resource_group: str, name: str, container_group: ContainerGroup) -> None:
-        """
-        Create a new container group
-
-        :param resource_group: the name of the resource group
-        :type resource_group: str
-        :param name: the name of the container group
-        :type name: str
-        :param container_group: the properties of the container group
-        :type container_group: azure.mgmt.containerinstance.models.ContainerGroup
-        """
-        self.connection.container_groups.create_or_update(resource_group, name, container_group)
-
-    def get_state_exitcode_details(self, resource_group: str, name: str) -> tuple:
-        """
-        Get the state and exitcode of a container group
-
-        :param resource_group: the name of the resource group
-        :type resource_group: str
-        :param name: the name of the container group
-        :type name: str
-        :return: A tuple with the state, exitcode, and details.
-            If the exitcode is unknown 0 is returned.
-        :rtype: tuple(state,exitcode,details)
-        """
-        warnings.warn(
-            "get_state_exitcode_details() is deprecated. Related method is get_state()",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        cg_state = self.get_state(resource_group, name)
-        c_state = cg_state.containers[0].instance_view.current_state
-        return (c_state.state, c_state.exit_code, c_state.detail_status)
-
-    def get_messages(self, resource_group: str, name: str) -> list:
-        """
-        Get the messages of a container group
-
-        :param resource_group: the name of the resource group
-        :type resource_group: str
-        :param name: the name of the container group
-        :type name: str
-        :return: A list of the event messages
-        :rtype: list[str]
-        """
-        warnings.warn(
-            "get_messages() is deprecated. Related method is get_state()", DeprecationWarning, stacklevel=2
-        )
-        cg_state = self.get_state(resource_group, name)
-        instance_view = cg_state.containers[0].instance_view
-        return [event.message for event in instance_view.events]
-
-    def get_state(self, resource_group: str, name: str) -> Any:
-        """
-        Get the state of a container group
-
-        :param resource_group: the name of the resource group
-        :type resource_group: str
-        :param name: the name of the container group
-        :type name: str
-        :return: ContainerGroup
-        :rtype: ~azure.mgmt.containerinstance.models.ContainerGroup
-        """
-        return self.connection.container_groups.get(resource_group, name, raw=False)
-
-    def get_logs(self, resource_group: str, name: str, tail: int = 1000) -> list:
-        """
-        Get the tail from logs of a container group
-
-        :param resource_group: the name of the resource group
-        :type resource_group: str
-        :param name: the name of the container group
-        :type name: str
-        :param tail: the size of the tail
-        :type tail: int
-        :return: A list of log messages
-        :rtype: list[str]
-        """
-        logs = self.connection.container.list_logs(resource_group, name, name, tail=tail)
-        return logs.content.splitlines(True)
-
-    def delete(self, resource_group: str, name: str) -> None:
-        """
-        Delete a container group
-
-        :param resource_group: the name of the resource group
-        :type resource_group: str
-        :param name: the name of the container group
-        :type name: str
-        """
-        self.connection.container_groups.delete(resource_group, name)
-
-    def exists(self, resource_group: str, name: str) -> bool:
-        """
-        Test if a container group exists
-
-        :param resource_group: the name of the resource group
-        :type resource_group: str
-        :param name: the name of the container group
-        :type name: str
-        """
-        for container in self.connection.container_groups.list_by_resource_group(resource_group):
-            if container.name == name:
-                return True
-        return False
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.microsoft.azure.hooks.container_instance`.",
+    DeprecationWarning,
+    stacklevel=2,
+)
