+++ airflow/providers/microsoft/azure/hooks/azure_container_registry.py
@@ -1,4 +1,3 @@
-#
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -15,52 +14,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""Hook for Azure Container Registry"""
+"""This module is deprecated. Please use :mod:`airflow.providers.microsoft.azure.hooks.container_registry`."""
 
-from typing import Dict
+import warnings
 
-from azure.mgmt.containerinstance.models import ImageRegistryCredential
+from airflow.providers.microsoft.azure.hooks.container_registry import AzureContainerRegistryHook  # noqa
 
-from airflow.hooks.base import BaseHook
-
-
-class AzureContainerRegistryHook(BaseHook):
-    """
-    A hook to communicate with a Azure Container Registry.
-
-    :param conn_id: :ref:`Azure Container Registry connection id<howto/connection:acr>`
-        of a service principal which will be used to start the container instance
-
-    :type conn_id: str
-    """
-
-    conn_name_attr = 'azure_container_registry_conn_id'
-    default_conn_name = 'azure_container_registry_default'
-    conn_type = 'azure_container_registry'
-    hook_name = 'Azure Container Registry'
-
-    @staticmethod
-    def get_ui_field_behaviour() -> Dict:
-        """Returns custom field behaviour"""
-        return {
-            "hidden_fields": ['schema', 'port', 'extra'],
-            "relabeling": {
-                'login': 'Registry Username',
-                'password': 'Registry Password',
-                'host': 'Registry Server',
-            },
-            "placeholders": {
-                'login': 'private registry username',
-                'password': 'private registry password',
-                'host': 'docker image registry server',
-            },
-        }
-
-    def __init__(self, conn_id: str = 'azure_registry') -> None:
-        super().__init__()
-        self.conn_id = conn_id
-        self.connection = self.get_conn()
-
-    def get_conn(self) -> ImageRegistryCredential:
-        conn = self.get_connection(self.conn_id)
-        return ImageRegistryCredential(server=conn.host, username=conn.login, password=conn.password)
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.microsoft.azure.hooks.container_registry`.",
+    DeprecationWarning,
+    stacklevel=2,
+)
