+++ airflow/providers/microsoft/azure/provider.yaml
@@ -92,12 +92,15 @@ operators:
       - airflow.providers.microsoft.azure.operators.adx
   - integration-name: Microsoft Azure Batch
     python-modules:
+      - airflow.providers.microsoft.azure.operators.batch
       - airflow.providers.microsoft.azure.operators.azure_batch
   - integration-name: Microsoft Azure Container Instances
     python-modules:
+      - airflow.providers.microsoft.azure.operators.container_instances
       - airflow.providers.microsoft.azure.operators.azure_container_instances
   - integration-name: Microsoft Azure Cosmos DB
     python-modules:
+      - airflow.providers.microsoft.azure.operators.cosmos
       - airflow.providers.microsoft.azure.operators.azure_cosmos
   - integration-name: Microsoft Azure Blob Storage
     python-modules:
@@ -109,6 +112,7 @@ operators:
 sensors:
   - integration-name: Microsoft Azure Cosmos DB
     python-modules:
+      - airflow.providers.microsoft.azure.sensors.cosmos
       - airflow.providers.microsoft.azure.sensors.azure_cosmos
   - integration-name: Microsoft Azure Blob Storage
     python-modules:
@@ -120,6 +124,9 @@ sensors:
 hooks:
   - integration-name: Microsoft Azure Container Instances
     python-modules:
+      - airflow.providers.microsoft.azure.hooks.container_volume
+      - airflow.providers.microsoft.azure.hooks.container_registry
+      - airflow.providers.microsoft.azure.hooks.container_instance
       - airflow.providers.microsoft.azure.hooks.azure_container_volume
       - airflow.providers.microsoft.azure.hooks.azure_container_registry
       - airflow.providers.microsoft.azure.hooks.azure_container_instance
@@ -128,18 +135,22 @@ hooks:
       - airflow.providers.microsoft.azure.hooks.adx
   - integration-name: Microsoft Azure FileShare
     python-modules:
+      - airflow.providers.microsoft.azure.hooks.fileshare
       - airflow.providers.microsoft.azure.hooks.azure_fileshare
   - integration-name: Microsoft Azure
     python-modules:
       - airflow.providers.microsoft.azure.hooks.base_azure
   - integration-name: Microsoft Azure Batch
     python-modules:
+      - airflow.providers.microsoft.azure.hooks.batch
       - airflow.providers.microsoft.azure.hooks.azure_batch
   - integration-name: Microsoft Azure Data Lake Storage
     python-modules:
+      - airflow.providers.microsoft.azure.hooks.data_lake
       - airflow.providers.microsoft.azure.hooks.azure_data_lake
   - integration-name: Microsoft Azure Cosmos DB
     python-modules:
+      - airflow.providers.microsoft.azure.hooks.cosmos
       - airflow.providers.microsoft.azure.hooks.azure_cosmos
   - integration-name: Microsoft Azure Blob Storage
     python-modules:
@@ -171,43 +182,44 @@ transfers:
 hook-class-names:  # deprecated - to be removed after providers add dependency on Airflow 2.2.0+
   - airflow.providers.microsoft.azure.hooks.base_azure.AzureBaseHook
   - airflow.providers.microsoft.azure.hooks.adx.AzureDataExplorerHook
-  - airflow.providers.microsoft.azure.hooks.azure_batch.AzureBatchHook
-  - airflow.providers.microsoft.azure.hooks.azure_cosmos.AzureCosmosDBHook
-  - airflow.providers.microsoft.azure.hooks.azure_data_lake.AzureDataLakeHook
-  - airflow.providers.microsoft.azure.hooks.azure_fileshare.AzureFileShareHook
-  - airflow.providers.microsoft.azure.hooks.azure_container_volume.AzureContainerVolumeHook
-  - airflow.providers.microsoft.azure.hooks.azure_container_instance.AzureContainerInstanceHook
+  - airflow.providers.microsoft.azure.hooks.batch.AzureBatchHook
+  - airflow.providers.microsoft.azure.hooks.cosmos.AzureCosmosDBHook
+  - airflow.providers.microsoft.azure.hooks.data_lake.AzureDataLakeHook
+  - airflow.providers.microsoft.azure.hooks.fileshare.AzureFileShareHook
+  - airflow.providers.microsoft.azure.hooks.container_volume.AzureContainerVolumeHook
+  - airflow.providers.microsoft.azure.hooks.container_instance.AzureContainerInstanceHook
   - airflow.providers.microsoft.azure.hooks.wasb.WasbHook
   - airflow.providers.microsoft.azure.hooks.data_factory.AzureDataFactoryHook
-  - airflow.providers.microsoft.azure.hooks.azure_container_registry.AzureContainerRegistryHook
+  - airflow.providers.microsoft.azure.hooks.container_registry.AzureContainerRegistryHook
 
 connection-types:
   - hook-class-name: airflow.providers.microsoft.azure.hooks.base_azure.AzureBaseHook
     connection-type: azure
   - hook-class-name: airflow.providers.microsoft.azure.hooks.adx.AzureDataExplorerHook
     connection-type: azure_data_explorer
-  - hook-class-name: airflow.providers.microsoft.azure.hooks.azure_batch.AzureBatchHook
+  - hook-class-name: airflow.providers.microsoft.azure.hooks.batch.AzureBatchHook
     connection-type: azure_batch
-  - hook-class-name: airflow.providers.microsoft.azure.hooks.azure_cosmos.AzureCosmosDBHook
+  - hook-class-name: airflow.providers.microsoft.azure.hooks.cosmos.AzureCosmosDBHook
     connection-type: azure_cosmos
-  - hook-class-name: airflow.providers.microsoft.azure.hooks.azure_data_lake.AzureDataLakeHook
+  - hook-class-name: airflow.providers.microsoft.azure.hooks.data_lake.AzureDataLakeHook
     connection-type: azure_data_lake
-  - hook-class-name: airflow.providers.microsoft.azure.hooks.azure_fileshare.AzureFileShareHook
+  - hook-class-name: airflow.providers.microsoft.azure.hooks.fileshare.AzureFileShareHook
     connection-type: azure_fileshare
-  - hook-class-name: airflow.providers.microsoft.azure.hooks.azure_container_volume.AzureContainerVolumeHook
+  - hook-class-name: airflow.providers.microsoft.azure.hooks.container_volume.AzureContainerVolumeHook
     connection-type: azure_container_volume
   - hook-class-name: >-
-      airflow.providers.microsoft.azure.hooks.azure_container_instance.AzureContainerInstanceHook
+      airflow.providers.microsoft.azure.hooks.container_instance.AzureContainerInstanceHook
     connection-type: azure_container_instance
   - hook-class-name: airflow.providers.microsoft.azure.hooks.wasb.WasbHook
     connection-type: wasb
   - hook-class-name: airflow.providers.microsoft.azure.hooks.data_factory.AzureDataFactoryHook
     connection-type: azure_data_factory
   - hook-class-name: >-
-      airflow.providers.microsoft.azure.hooks.azure_container_registry.AzureContainerRegistryHook
+      airflow.providers.microsoft.azure.hooks.container_registry.AzureContainerRegistryHook
     connection-type: azure_container_registry
 
 secrets-backends:
+  - airflow.providers.microsoft.azure.secrets.key_vault.AzureKeyVaultBackend
   - airflow.providers.microsoft.azure.secrets.azure_key_vault.AzureKeyVaultBackend
 
 logging:
