+++ airflow/providers/microsoft/azure/sensors/azure_cosmos.py
@@ -1,4 +1,3 @@
-#
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -15,54 +14,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+"""This module is deprecated. Please use :mod:`airflow.providers.microsoft.azure.sensors.cosmos`."""
 
-from airflow.providers.microsoft.azure.hooks.azure_cosmos import AzureCosmosDBHook
-from airflow.sensors.base import BaseSensorOperator
+import warnings
 
+from airflow.providers.microsoft.azure.sensors.cosmos import AzureCosmosDocumentSensor  # noqa
 
-class AzureCosmosDocumentSensor(BaseSensorOperator):
-    """
-    Checks for the existence of a document which
-    matches the given query in CosmosDB. Example:
-
-    .. code-block::
-
-    azure_cosmos_sensor = AzureCosmosDocumentSensor(
-        database_name="somedatabase_name",
-        collection_name="somecollection_name",
-        document_id="unique-doc-id",
-        azure_cosmos_conn_id="azure_cosmos_default",
-        task_id="azure_cosmos_sensor")
-
-    :param database_name: Target CosmosDB database_name.
-    :type database_name: str
-    :param collection_name: Target CosmosDB collection_name.
-    :type collection_name: str
-    :param document_id: The ID of the target document.
-    :type document_id: str
-    :param azure_cosmos_conn_id: Reference to the
-        :ref:`Azure CosmosDB connection<howto/connection:azure_cosmos>`.
-    :type azure_cosmos_conn_id: str
-    """
-
-    template_fields = ('database_name', 'collection_name', 'document_id')
-
-    def __init__(
-        self,
-        *,
-        database_name: str,
-        collection_name: str,
-        document_id: str,
-        azure_cosmos_conn_id: str = "azure_cosmos_default",
-        **kwargs,
-    ) -> None:
-        super().__init__(**kwargs)
-        self.azure_cosmos_conn_id = azure_cosmos_conn_id
-        self.database_name = database_name
-        self.collection_name = collection_name
-        self.document_id = document_id
-
-    def poke(self, context: dict) -> bool:
-        self.log.info("*** Intering poke")
-        hook = AzureCosmosDBHook(self.azure_cosmos_conn_id)
-        return hook.get_document(self.document_id, self.database_name, self.collection_name) is not None
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.microsoft.azure.sensors.cosmos`.",
+    DeprecationWarning,
+    stacklevel=2,
+)
