+++ airflow/providers/microsoft/azure/transfers/local_to_adls.py
@@ -19,7 +19,7 @@ from typing import Any, Dict, Optional
 
 from airflow.exceptions import AirflowException
 from airflow.models import BaseOperator
-from airflow.providers.microsoft.azure.hooks.azure_data_lake import AzureDataLakeHook
+from airflow.providers.microsoft.azure.hooks.data_lake import AzureDataLakeHook
 
 
 class LocalFilesystemToADLSOperator(BaseOperator):
