+++ airflow/providers/microsoft/azure/transfers/oracle_to_azure_data_lake.py
@@ -23,7 +23,7 @@ from typing import Any, Optional, Union
 import unicodecsv as csv
 
 from airflow.models import BaseOperator
-from airflow.providers.microsoft.azure.hooks.azure_data_lake import AzureDataLakeHook
+from airflow.providers.microsoft.azure.hooks.data_lake import AzureDataLakeHook
 from airflow.providers.oracle.hooks.oracle import OracleHook
