+++ tests/providers/microsoft/azure/hooks/test_azure_container_instance.py
@@ -29,7 +29,7 @@ from azure.mgmt.containerinstance.models import (
 )
 
 from airflow.models import Connection
-from airflow.providers.microsoft.azure.hooks.azure_container_instance import AzureContainerInstanceHook
+from airflow.providers.microsoft.azure.hooks.container_instance import AzureContainerInstanceHook
 from airflow.utils import db
