+++ tests/providers/microsoft/azure/hooks/test_azure_cosmos.py
@@ -29,7 +29,7 @@ from azure.cosmos.cosmos_client import CosmosClient
 
 from airflow.exceptions import AirflowException
 from airflow.models import Connection
-from airflow.providers.microsoft.azure.hooks.azure_cosmos import AzureCosmosDBHook
+from airflow.providers.microsoft.azure.hooks.cosmos import AzureCosmosDBHook
 from airflow.utils import db
 
 
@@ -59,13 +59,13 @@ class TestAzureCosmosDbHook(unittest.TestCase):
             )
         )
 
-    @mock.patch('airflow.providers.microsoft.azure.hooks.azure_cosmos.CosmosClient', autospec=True)
+    @mock.patch('airflow.providers.microsoft.azure.hooks.cosmos.CosmosClient', autospec=True)
     def test_client(self, mock_cosmos):
         hook = AzureCosmosDBHook(azure_cosmos_conn_id='azure_cosmos_test_key_id')
         assert hook._conn is None
         assert isinstance(hook.get_conn(), CosmosClient)
 
-    @mock.patch('airflow.providers.microsoft.azure.hooks.azure_cosmos.CosmosClient')
+    @mock.patch('airflow.providers.microsoft.azure.hooks.cosmos.CosmosClient')
     def test_create_database(self, mock_cosmos):
         hook = AzureCosmosDBHook(azure_cosmos_conn_id='azure_cosmos_test_key_id')
         hook.create_database(self.test_database_name)
@@ -73,19 +73,19 @@ class TestAzureCosmosDbHook(unittest.TestCase):
         mock_cosmos.assert_any_call(self.test_end_point, {'masterKey': self.test_master_key})
         mock_cosmos.assert_has_calls(expected_calls)
 
-    @mock.patch('airflow.providers.microsoft.azure.hooks.azure_cosmos.CosmosClient')
+    @mock.patch('airflow.providers.microsoft.azure.hooks.cosmos.CosmosClient')
     def test_create_database_exception(self, mock_cosmos):
         hook = AzureCosmosDBHook(azure_cosmos_conn_id='azure_cosmos_test_key_id')
         with pytest.raises(AirflowException):
             hook.create_database(None)
 
-    @mock.patch('airflow.providers.microsoft.azure.hooks.azure_cosmos.CosmosClient')
+    @mock.patch('airflow.providers.microsoft.azure.hooks.cosmos.CosmosClient')
     def test_create_container_exception(self, mock_cosmos):
         hook = AzureCosmosDBHook(azure_cosmos_conn_id='azure_cosmos_test_key_id')
         with pytest.raises(AirflowException):
             hook.create_collection(None)
 
-    @mock.patch('airflow.providers.microsoft.azure.hooks.azure_cosmos.CosmosClient')
+    @mock.patch('airflow.providers.microsoft.azure.hooks.cosmos.CosmosClient')
     def test_create_container(self, mock_cosmos):
         hook = AzureCosmosDBHook(azure_cosmos_conn_id='azure_cosmos_test_key_id')
         hook.create_collection(self.test_collection_name, self.test_database_name)
@@ -95,7 +95,7 @@ class TestAzureCosmosDbHook(unittest.TestCase):
         mock_cosmos.assert_any_call(self.test_end_point, {'masterKey': self.test_master_key})
         mock_cosmos.assert_has_calls(expected_calls)
 
-    @mock.patch('airflow.providers.microsoft.azure.hooks.azure_cosmos.CosmosClient')
+    @mock.patch('airflow.providers.microsoft.azure.hooks.cosmos.CosmosClient')
     def test_create_container_default(self, mock_cosmos):
         hook = AzureCosmosDBHook(azure_cosmos_conn_id='azure_cosmos_test_key_id')
         hook.create_collection(self.test_collection_name)
@@ -105,7 +105,7 @@ class TestAzureCosmosDbHook(unittest.TestCase):
         mock_cosmos.assert_any_call(self.test_end_point, {'masterKey': self.test_master_key})
         mock_cosmos.assert_has_calls(expected_calls)
 
-    @mock.patch('airflow.providers.microsoft.azure.hooks.azure_cosmos.CosmosClient')
+    @mock.patch('airflow.providers.microsoft.azure.hooks.cosmos.CosmosClient')
     def test_upsert_document_default(self, mock_cosmos):
         test_id = str(uuid.uuid4())
         mock_cosmos.return_value.CreateItem.return_value = {'id': test_id}
@@ -122,7 +122,7 @@ class TestAzureCosmosDbHook(unittest.TestCase):
         logging.getLogger().info(returned_item)
         assert returned_item['id'] == test_id
 
-    @mock.patch('airflow.providers.microsoft.azure.hooks.azure_cosmos.CosmosClient')
+    @mock.patch('airflow.providers.microsoft.azure.hooks.cosmos.CosmosClient')
     def test_upsert_document(self, mock_cosmos):
         test_id = str(uuid.uuid4())
         mock_cosmos.return_value.CreateItem.return_value = {'id': test_id}
@@ -146,7 +146,7 @@ class TestAzureCosmosDbHook(unittest.TestCase):
         logging.getLogger().info(returned_item)
         assert returned_item['id'] == test_id
 
-    @mock.patch('airflow.providers.microsoft.azure.hooks.azure_cosmos.CosmosClient')
+    @mock.patch('airflow.providers.microsoft.azure.hooks.cosmos.CosmosClient')
     def test_insert_documents(self, mock_cosmos):
         test_id1 = str(uuid.uuid4())
         test_id2 = str(uuid.uuid4())
@@ -177,7 +177,7 @@ class TestAzureCosmosDbHook(unittest.TestCase):
         mock_cosmos.assert_any_call(self.test_end_point, {'masterKey': self.test_master_key})
         mock_cosmos.assert_has_calls(expected_calls, any_order=True)
 
-    @mock.patch('airflow.providers.microsoft.azure.hooks.azure_cosmos.CosmosClient')
+    @mock.patch('airflow.providers.microsoft.azure.hooks.cosmos.CosmosClient')
     def test_delete_database(self, mock_cosmos):
         hook = AzureCosmosDBHook(azure_cosmos_conn_id='azure_cosmos_test_key_id')
         hook.delete_database(self.test_database_name)
@@ -185,7 +185,7 @@ class TestAzureCosmosDbHook(unittest.TestCase):
         mock_cosmos.assert_any_call(self.test_end_point, {'masterKey': self.test_master_key})
         mock_cosmos.assert_has_calls(expected_calls)
 
-    @mock.patch('airflow.providers.microsoft.azure.hooks.azure_cosmos.CosmosClient')
+    @mock.patch('airflow.providers.microsoft.azure.hooks.cosmos.CosmosClient')
     def test_delete_database_exception(self, mock_cosmos):
         hook = AzureCosmosDBHook(azure_cosmos_conn_id='azure_cosmos_test_key_id')
         with pytest.raises(AirflowException):
@@ -197,7 +197,7 @@ class TestAzureCosmosDbHook(unittest.TestCase):
         with pytest.raises(AirflowException):
             hook.delete_collection(None)
 
-    @mock.patch('airflow.providers.microsoft.azure.hooks.azure_cosmos.CosmosClient')
+    @mock.patch('airflow.providers.microsoft.azure.hooks.cosmos.CosmosClient')
     def test_delete_container(self, mock_cosmos):
         hook = AzureCosmosDBHook(azure_cosmos_conn_id='azure_cosmos_test_key_id')
         hook.delete_collection(self.test_collection_name, self.test_database_name)
@@ -205,7 +205,7 @@ class TestAzureCosmosDbHook(unittest.TestCase):
         mock_cosmos.assert_any_call(self.test_end_point, {'masterKey': self.test_master_key})
         mock_cosmos.assert_has_calls(expected_calls)
 
-    @mock.patch('airflow.providers.microsoft.azure.hooks.azure_cosmos.CosmosClient')
+    @mock.patch('airflow.providers.microsoft.azure.hooks.cosmos.CosmosClient')
     def test_delete_container_default(self, mock_cosmos):
         hook = AzureCosmosDBHook(azure_cosmos_conn_id='azure_cosmos_test_key_id')
         hook.delete_collection(self.test_collection_name)
