+++ tests/providers/microsoft/azure/hooks/test_azure_data_lake.py
@@ -36,11 +36,11 @@ class TestAzureDataLakeHook(unittest.TestCase):
             )
         )
 
-    @mock.patch('airflow.providers.microsoft.azure.hooks.azure_data_lake.lib', autospec=True)
+    @mock.patch('airflow.providers.microsoft.azure.hooks.data_lake.lib', autospec=True)
     def test_conn(self, mock_lib):
         from azure.datalake.store import core
 
-        from airflow.providers.microsoft.azure.hooks.azure_data_lake import AzureDataLakeHook
+        from airflow.providers.microsoft.azure.hooks.data_lake import AzureDataLakeHook
 
         hook = AzureDataLakeHook(azure_data_lake_conn_id='adl_test_key')
         assert hook._conn is None
@@ -48,23 +48,19 @@ class TestAzureDataLakeHook(unittest.TestCase):
         assert isinstance(hook.get_conn(), core.AzureDLFileSystem)
         assert mock_lib.auth.called
 
-    @mock.patch(
-        'airflow.providers.microsoft.azure.hooks.azure_data_lake.core.AzureDLFileSystem', autospec=True
-    )
-    @mock.patch('airflow.providers.microsoft.azure.hooks.azure_data_lake.lib', autospec=True)
+    @mock.patch('airflow.providers.microsoft.azure.hooks.data_lake.core.AzureDLFileSystem', autospec=True)
+    @mock.patch('airflow.providers.microsoft.azure.hooks.data_lake.lib', autospec=True)
     def test_check_for_blob(self, mock_lib, mock_filesystem):
-        from airflow.providers.microsoft.azure.hooks.azure_data_lake import AzureDataLakeHook
+        from airflow.providers.microsoft.azure.hooks.data_lake import AzureDataLakeHook
 
         hook = AzureDataLakeHook(azure_data_lake_conn_id='adl_test_key')
         hook.check_for_file('file_path')
         mock_filesystem.glob.called
 
-    @mock.patch(
-        'airflow.providers.microsoft.azure.hooks.azure_data_lake.multithread.ADLUploader', autospec=True
-    )
-    @mock.patch('airflow.providers.microsoft.azure.hooks.azure_data_lake.lib', autospec=True)
+    @mock.patch('airflow.providers.microsoft.azure.hooks.data_lake.multithread.ADLUploader', autospec=True)
+    @mock.patch('airflow.providers.microsoft.azure.hooks.data_lake.lib', autospec=True)
     def test_upload_file(self, mock_lib, mock_uploader):
-        from airflow.providers.microsoft.azure.hooks.azure_data_lake import AzureDataLakeHook
+        from airflow.providers.microsoft.azure.hooks.data_lake import AzureDataLakeHook
 
         hook = AzureDataLakeHook(azure_data_lake_conn_id='adl_test_key')
         hook.upload_file(
@@ -85,12 +81,10 @@ class TestAzureDataLakeHook(unittest.TestCase):
             blocksize=4194304,
         )
 
-    @mock.patch(
-        'airflow.providers.microsoft.azure.hooks.azure_data_lake.multithread.ADLDownloader', autospec=True
-    )
-    @mock.patch('airflow.providers.microsoft.azure.hooks.azure_data_lake.lib', autospec=True)
+    @mock.patch('airflow.providers.microsoft.azure.hooks.data_lake.multithread.ADLDownloader', autospec=True)
+    @mock.patch('airflow.providers.microsoft.azure.hooks.data_lake.lib', autospec=True)
     def test_download_file(self, mock_lib, mock_downloader):
-        from airflow.providers.microsoft.azure.hooks.azure_data_lake import AzureDataLakeHook
+        from airflow.providers.microsoft.azure.hooks.data_lake import AzureDataLakeHook
 
         hook = AzureDataLakeHook(azure_data_lake_conn_id='adl_test_key')
         hook.download_file(
@@ -111,34 +105,28 @@ class TestAzureDataLakeHook(unittest.TestCase):
             blocksize=4194304,
         )
 
-    @mock.patch(
-        'airflow.providers.microsoft.azure.hooks.azure_data_lake.core.AzureDLFileSystem', autospec=True
-    )
-    @mock.patch('airflow.providers.microsoft.azure.hooks.azure_data_lake.lib', autospec=True)
+    @mock.patch('airflow.providers.microsoft.azure.hooks.data_lake.core.AzureDLFileSystem', autospec=True)
+    @mock.patch('airflow.providers.microsoft.azure.hooks.data_lake.lib', autospec=True)
     def test_list_glob(self, mock_lib, mock_fs):
-        from airflow.providers.microsoft.azure.hooks.azure_data_lake import AzureDataLakeHook
+        from airflow.providers.microsoft.azure.hooks.data_lake import AzureDataLakeHook
 
         hook = AzureDataLakeHook(azure_data_lake_conn_id='adl_test_key')
         hook.list('file_path/*')
         mock_fs.return_value.glob.assert_called_once_with('file_path/*')
 
-    @mock.patch(
-        'airflow.providers.microsoft.azure.hooks.azure_data_lake.core.AzureDLFileSystem', autospec=True
-    )
-    @mock.patch('airflow.providers.microsoft.azure.hooks.azure_data_lake.lib', autospec=True)
+    @mock.patch('airflow.providers.microsoft.azure.hooks.data_lake.core.AzureDLFileSystem', autospec=True)
+    @mock.patch('airflow.providers.microsoft.azure.hooks.data_lake.lib', autospec=True)
     def test_list_walk(self, mock_lib, mock_fs):
-        from airflow.providers.microsoft.azure.hooks.azure_data_lake import AzureDataLakeHook
+        from airflow.providers.microsoft.azure.hooks.data_lake import AzureDataLakeHook
 
         hook = AzureDataLakeHook(azure_data_lake_conn_id='adl_test_key')
         hook.list('file_path/some_folder/')
         mock_fs.return_value.walk.assert_called_once_with('file_path/some_folder/')
 
-    @mock.patch(
-        'airflow.providers.microsoft.azure.hooks.azure_data_lake.core.AzureDLFileSystem', autospec=True
-    )
-    @mock.patch('airflow.providers.microsoft.azure.hooks.azure_data_lake.lib', autospec=True)
+    @mock.patch('airflow.providers.microsoft.azure.hooks.data_lake.core.AzureDLFileSystem', autospec=True)
+    @mock.patch('airflow.providers.microsoft.azure.hooks.data_lake.lib', autospec=True)
     def test_remove(self, mock_lib, mock_fs):
-        from airflow.providers.microsoft.azure.hooks.azure_data_lake import AzureDataLakeHook
+        from airflow.providers.microsoft.azure.hooks.data_lake import AzureDataLakeHook
 
         hook = AzureDataLakeHook(azure_data_lake_conn_id='adl_test_key')
         hook.remove('filepath', True)
