+++ tests/providers/microsoft/azure/secrets/test_azure_key_vault.py
@@ -20,19 +20,19 @@ from unittest import TestCase, mock
 
 from azure.core.exceptions import ResourceNotFoundError
 
-from airflow.providers.microsoft.azure.secrets.azure_key_vault import AzureKeyVaultBackend
+from airflow.providers.microsoft.azure.secrets.key_vault import AzureKeyVaultBackend
 
 
 class TestAzureKeyVaultBackend(TestCase):
-    @mock.patch('airflow.providers.microsoft.azure.secrets.azure_key_vault.AzureKeyVaultBackend.get_conn_uri')
+    @mock.patch('airflow.providers.microsoft.azure.secrets.key_vault.AzureKeyVaultBackend.get_conn_uri')
     def test_get_connections(self, mock_get_uri):
         mock_get_uri.return_value = 'scheme://user:pass@host:100'
         conn_list = AzureKeyVaultBackend().get_connections('fake_conn')
         conn = conn_list[0]
         assert conn.host == 'host'
 
-    @mock.patch('airflow.providers.microsoft.azure.secrets.azure_key_vault.DefaultAzureCredential')
-    @mock.patch('airflow.providers.microsoft.azure.secrets.azure_key_vault.SecretClient')
+    @mock.patch('airflow.providers.microsoft.azure.secrets.key_vault.DefaultAzureCredential')
+    @mock.patch('airflow.providers.microsoft.azure.secrets.key_vault.SecretClient')
     def test_get_conn_uri(self, mock_secret_client, mock_azure_cred):
         mock_cred = mock.Mock()
         mock_sec_client = mock.Mock()
@@ -50,7 +50,7 @@ class TestAzureKeyVaultBackend(TestCase):
         )
         assert returned_uri == 'postgresql://airflow:airflow@host:5432/airflow'
 
-    @mock.patch('airflow.providers.microsoft.azure.secrets.azure_key_vault.AzureKeyVaultBackend.client')
+    @mock.patch('airflow.providers.microsoft.azure.secrets.key_vault.AzureKeyVaultBackend.client')
     def test_get_conn_uri_non_existent_key(self, mock_client):
         """
         Test that if the key with connection ID is not present,
@@ -63,7 +63,7 @@ class TestAzureKeyVaultBackend(TestCase):
         assert backend.get_conn_uri(conn_id=conn_id) is None
         assert [] == backend.get_connections(conn_id=conn_id)
 
-    @mock.patch('airflow.providers.microsoft.azure.secrets.azure_key_vault.AzureKeyVaultBackend.client')
+    @mock.patch('airflow.providers.microsoft.azure.secrets.key_vault.AzureKeyVaultBackend.client')
     def test_get_variable(self, mock_client):
         mock_client.get_secret.return_value = mock.Mock(value='world')
         backend = AzureKeyVaultBackend()
@@ -71,7 +71,7 @@ class TestAzureKeyVaultBackend(TestCase):
         mock_client.get_secret.assert_called_with(name='airflow-variables-hello')
         assert 'world' == returned_uri
 
-    @mock.patch('airflow.providers.microsoft.azure.secrets.azure_key_vault.AzureKeyVaultBackend.client')
+    @mock.patch('airflow.providers.microsoft.azure.secrets.key_vault.AzureKeyVaultBackend.client')
     def test_get_variable_non_existent_key(self, mock_client):
         """
         Test that if Variable key is not present,
@@ -81,7 +81,7 @@ class TestAzureKeyVaultBackend(TestCase):
         backend = AzureKeyVaultBackend()
         assert backend.get_variable('test_mysql') is None
 
-    @mock.patch('airflow.providers.microsoft.azure.secrets.azure_key_vault.AzureKeyVaultBackend.client')
+    @mock.patch('airflow.providers.microsoft.azure.secrets.key_vault.AzureKeyVaultBackend.client')
     def test_get_secret_value_not_found(self, mock_client):
         """
         Test that if a non-existent secret returns None
@@ -92,7 +92,7 @@ class TestAzureKeyVaultBackend(TestCase):
             backend._get_secret(path_prefix=backend.connections_prefix, secret_id='test_non_existent') is None
         )
 
-    @mock.patch('airflow.providers.microsoft.azure.secrets.azure_key_vault.AzureKeyVaultBackend.client')
+    @mock.patch('airflow.providers.microsoft.azure.secrets.key_vault.AzureKeyVaultBackend.client')
     def test_get_secret_value(self, mock_client):
         """
         Test that get_secret returns the secret value
@@ -103,7 +103,7 @@ class TestAzureKeyVaultBackend(TestCase):
         mock_client.get_secret.assert_called_with(name='af-secrets-test-mysql-password')
         assert secret_val == 'super-secret'
 
-    @mock.patch('airflow.providers.microsoft.azure.secrets.azure_key_vault.AzureKeyVaultBackend._get_secret')
+    @mock.patch('airflow.providers.microsoft.azure.secrets.key_vault.AzureKeyVaultBackend._get_secret')
     def test_connection_prefix_none_value(self, mock_get_secret):
         """
         Test that if Connections prefix is None,
@@ -116,7 +116,7 @@ class TestAzureKeyVaultBackend(TestCase):
         assert backend.get_conn_uri('test_mysql') is None
         mock_get_secret.assert_not_called()
 
-    @mock.patch('airflow.providers.microsoft.azure.secrets.azure_key_vault.AzureKeyVaultBackend._get_secret')
+    @mock.patch('airflow.providers.microsoft.azure.secrets.key_vault.AzureKeyVaultBackend._get_secret')
     def test_variable_prefix_none_value(self, mock_get_secret):
         """
         Test that if Variables prefix is None,
@@ -129,7 +129,7 @@ class TestAzureKeyVaultBackend(TestCase):
         assert backend.get_variable('hello') is None
         mock_get_secret.assert_not_called()
 
-    @mock.patch('airflow.providers.microsoft.azure.secrets.azure_key_vault.AzureKeyVaultBackend._get_secret')
+    @mock.patch('airflow.providers.microsoft.azure.secrets.key_vault.AzureKeyVaultBackend._get_secret')
     def test_config_prefix_none_value(self, mock_get_secret):
         """
         Test that if Config prefix is None,
